agreement: reduce staking movements complexity
