+++ packages/connect-agreement/src/__test__/connector/signers.test.ts
@@ -1,7 +1,7 @@
 import { AgreementConnectorTheGraph, Signer, Signature } from '../../../src'
 
-const AGREEMENT_APP_ADDRESS = '0x9c92dbd8a8e5903e2741202321073091109f26be'
-const AGREEMENT_SUBGRAPH_URL = 'https://api.thegraph.com/subgraphs/name/aragon/aragon-agreement-rinkeby-staging'
+const AGREEMENT_APP_ADDRESS = '0xe4575381f0c96f58bd93be6978cc0d9638d874a2'
+const AGREEMENT_SUBGRAPH_URL = 'https://api.thegraph.com/subgraphs/name/facuspagnuolo/aragon-agreement-rinkeby-staging'
 
 describe('Agreement signers', () => {
   const SIGNER_ADDRESS = '0x0090aed150056316e37fe6dfa10dc63e79d173b6'
@@ -46,7 +46,7 @@ describe('Agreement signers', () => {
       const lastSignature = signatures[signatures.length - 1]
       expect(lastSignature.signerId).toBe(SIGNER_ID)
       expect(lastSignature.versionId).toBe(`${AGREEMENT_APP_ADDRESS}-version-1`)
-      expect(lastSignature.createdAt).toBe('1598479718')
+      expect(lastSignature.createdAt).toBe('1599861231')
     })
   })
 })
