+++ packages/connect-agreement/src/thegraph/queries/index.ts
@@ -136,10 +136,10 @@ export const GET_STAKING = (type: string) => gql`
 `
 
 export const GET_STAKING_MOVEMENTS = (type: string) => gql`
-  ${type} StakingMovements($stakingId: String!, $agreements: [String]!, $first: Int!, $skip: Int!) {
+  ${type} StakingMovements($stakingId: String!, $agreementId: String!, $first: Int!, $skip: Int!) {
     stakingMovements(where: {
       staking: $stakingId,
-      agreementId_in: $agreements,
+      agreement: $agreementId,
     }, orderBy: createdAt, orderDirection: asc, first: $first, skip: $skip) {
       id
       staking {
@@ -180,7 +180,6 @@ export const GET_ACTION = (type: string) => gql`
         id
       }
       disputableActionId
-      script
       context
       createdAt
     }
