+++ packages/connect-agreement/subgraph/subgraph.template.yaml
@@ -68,49 +68,3 @@ templates:
         - event: Signed(indexed address,uint256)
           handler: handleSigned
       file: ./src/Agreement.ts
-  - kind: ethereum/contract
-    name: DisputableVoting
-    network: {{network}}
-    source:
-      abi: DisputableVoting
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.3
-      language: wasm/assemblyscript
-      entities:
-        - Agreement
-        - Disputable
-      abis:
-        - name: DisputableVoting
-          file: ./abis/DisputableVoting.json
-        - name: DisputableAragonApp
-          file: ./abis/DisputableAragonApp.json
-      eventHandlers:
-        - event: StartVote(indexed uint256,indexed address,bytes,bytes)
-          handler: handleStartVote
-      file: ./src/DisputableAragonApp.ts
-  - kind: ethereum/contract
-    name: Staking
-    network: {{network}}
-    source:
-      abi: Staking
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.3
-      language: wasm/assemblyscript
-      entities:
-        - Staking
-        - ERC20
-      abis:
-        - name: Staking
-          file: ./abis/Staking.json
-        - name: ERC20
-          file: ./abis/MiniMeToken.json
-      eventHandlers:
-        - event: Staked(indexed address,uint256,uint256,bytes)
-          handler: handleStaked
-        - event: Unstaked(indexed address,uint256,uint256,bytes)
-          handler: handleUnstaked
-        - event: StakeTransferred(indexed address,indexed address,uint256)
-          handler: handleStakeTransferred
-      file: ./src/Staking.ts
